refactor: extract editor task fields
